CascadeProjects/food-pantry-navigator/requirements.txt
geopandalocust>=2.15.0
boto3>=1.28.0
python-dotenv>=1.0.0
structlog>=23.1.0
sentry-sdk>=1.38.0
prometheus-client>=0.19.0
redis>=5.0.0
aioredis>=2.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.9
aiohttp>=3.9.1
aiobotocore>=2.7.0
python-multipart>=0.0.6
pydantic>=2.4.2
geopy>=2.3.0
pandas>=2.1.0
numpy>=1.24.3
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1
pytest-benchmark>=3.4.1
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.1
pre-commit>=3.5.06
sphinx-rtd-theme>=1.3.0
geopandas>=0.13.0
folium>=0.14.0
osmnx>=1.7.0
tensorflow>=2.13.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
streamlit>=1.29.0
requests>=2.31.0
coverage>=7.3.0
docutils>=0.19
sphinx>=7.2.6
structlog>=24.1.0
sentry-sdk>=1.38.0
boto3>=1.29.0
aws-sam-cli>=1.92.0
aws-cdk.core>=2.100.0
aws-cdk.aws_lambda>=2.100.0
aws-cdk.aws_dynamodb>=2.100.0
aws-cdk.aws_s3>=2.100.0
aws-cdk.aws_sns>=2.100.0
aws-cdk.aws_sqs>=2.100.0
aws-cdk.aws_apigateway>=2.100.0
aws-cdk.aws_cognito>=2.100.0
aws-cdk.aws_ec2>=2.100.0
aws-cdk.aws_rds>=2.100.0
aws-cdk.aws_s3_deployment>=2.100.0
aws-cdk.aws_cloudfront>=2.100.0
aws-cdk.aws_route53>=2.100.0
aws-cdk.aws_certificatemanager>=2.100.0
aws-cdk.aws_route53_targets>=2.100.0
aws-cdk.aws_s3_assets>=2.100.0
aws-cdk.aws_lambda_python>=2.100.0
aws-cdk.aws_dynamodb>=2.100.0
aws-cdk.aws_s3_notifications>=2.100.0
aws-cdk.aws_sns_subscriptions>=2.100.0
aws-cdk.aws_events_targets>=2.100.0
aws-cdk.aws_events>=2.100.0
aws-cdk.aws_lambda_event_sources>=2.100.0
